It's already possible to buy individual products. Now, click 'Pricing'. The real point of the site is to get a subscription that'll send your awesome sheep awesome shearing accessories monthly. After vigorous meetings with our investors, we've decided to offer 2 plans: the Farmer Brent at {dollar_sign}99/month and the New Zealander at {dollar_sign}199/month. But other than this fancy looking page, we haven't coded anything for this yet.
Tell Stripe about the Plans

Our first step is to tell Stripe about these two plans. To do that, open your trusty Stripe dashboard, make sure you're in the test environment, and find 'Plans' on the left. We're going to create the two plans in Stripe by hand. The ID can be any unique string and we'll use this forever after to refer to the plan. Use, farmer_brent_monthly. Then, fill in a name - that's less important and set the price as {dollar_sign}99 per month. Create that plan!

And yes, later, we'll add yearly versions of each plan... but one step at a time! Repeat that whole process for the New Zealander: set its ID to new_zealander_monthly, give it a name and set its price and interval. Perfect!