Now back to the controller! In the first Stripe tutorial, we worked with a ShoppingCart class that I created for us... because it's not really that important. It basically allows you to store products and a subscription in the user's session, so that as they surf around, we know what they have in their cart.
But before we use that, first get an instance of our new SubscriptionHelper object with {dollar_sign}subscriptionHelper = {dollar_sign}this->get('subscription_helper').
I already registered this as a service in Symfony:

 19 lines  app/config/services.yml
... lines 1 - 5
services:
... lines 7 - 15
  subscription_helper:
    class: AppBundle\Subscription\SubscriptionHelper
    autowire: true
Next, add {dollar_sign}plan = {dollar_sign}subscriptionHelper->findPlan() and pass it the {dollar_sign}planId

So this is nice: we give it the plan ID, and it gives us the corresponding, wonderful, SubscriptionPlan object:

 38 lines  src/AppBundle/Subscription/SubscriptionHelper.php
... lines 1 - 4
class SubscriptionHelper
{
... lines 7 - 24
  /**
  * @param
  * @return SubscriptionPlan|null
  */
  public function findPlan({dollar_sign}planId)
  {
    foreach ({dollar_sign}this->plans as {dollar_sign}plan) {
      if ({dollar_sign}plan->getPlanId() == {dollar_sign}planId) {
        return {dollar_sign}plan;
      }
    }
  }
}
But if the {dollar_sign}planId doesn't exist for some reason, throw {dollar_sign}this->createNotFoundException() to cause a 404 page.

Finally, add the plan to the cart, with {dollar_sign}this->get('shopping_cart')->addSubscription() and pass it the plan ID.

And boom! Our cart knows about the subscription! Finally, send them to the checkout page with return {dollar_sign}this->redirectToRoute('order_checkout') - that's the name of our checkoutAction route.

Okay team, give it a try! Add the Farmer Brent plan. Bah! We need to login: use the pre-filled email and the password used by all sheep: breakingbaad.

Ok, this looks kinda right: the total is {dollar_sign}99 because the ShoppingCart object knows about the subscription... but we haven't printed anything about the subscription in the cart table. So it looks weird.