Open up the Stripe docs and go down the page until you find subscriptions. There's a nice little 'Getting Started' section but the detailed guide is the place to go if you've got serious questions.
But let's start with the basics! Step 1... done! Step 2: subscribing your customers. Apparently all we need to do is set the plan on the Customer and save! Cool!

The Players: Subscription, Customer, Invoice

But actually, there's more going on behind the scenes. In reality, this will create a new object in Stripe: a Subscription. And actually, we're going to subscribe a user with slightly different code than this.

Keep reading below, the docs describe the lifecycle of a Subscription. For now, there's one really important thing to notice: when you create a Subscription, Stripe automatically creates an Invoice and charges that invoice immediately.

Open up the Stripe API docs so we can look at all the important objects so far. From part 1 of the tutorial, when someone buys individual products, we do a few things: we create or fetch a Customer, we create an InvoiceItem for each product and finally we create an Invoice and pay it. When you create an Invoice, Stripe automatically adds all unpaid invoice items to it.

With a Subscription, there are two new players: Plans and Subscriptions. Click 'Subscriptions' and go down to 'Create a Subscription'. Ah, so simple: a Subscription is between a Customer and a specific Plan. This is the code we will use.

Coding up the Stripe Subscription

Back on our site, after we fill out the checkout form, the whole thing submits to OrderController::checkoutAction(). And this passes the submitted Stripe token to chargeCustomer():

 108 lines  src/AppBundle/Controller/OrderController.php
... lines 1 - 11
class OrderController extends BaseController
{
... lines 14 - 48
  public function checkoutAction(Request {dollar_sign}request)
  {
... lines 51 - 53
    if ({dollar_sign}request->isMethod('POST')) {
... lines 55 - 56
      try {
        {dollar_sign}this->chargeCustomer({dollar_sign}token);
      } catch (\Stripe\Error\Card {dollar_sign}e) {
        {dollar_sign}error = 'There was a problem charging your card: '.{dollar_sign}e->getMessage();
      }
... lines 62 - 68
        }
... lines 70 - 77
    }
... lines 79 - 105
}
... lines 107 - 108
Ah that's where the magic happens: it creates or gets the Customer, adds InvoiceItems and creates the Invoice:

 108 lines  src/AppBundle/Controller/OrderController.php
... lines 1 - 11
class OrderController extends BaseController
{
... lines 14 - 83
  private function chargeCustomer({dollar_sign}token)
  {
    {dollar_sign}stripeClient = {dollar_sign}this->get('stripe_client');
    /** @var User {dollar_sign}user */
    {dollar_sign}user = {dollar_sign}this->getUser();
    if (! {dollar_sign}user->getStripeCustomerId()) {
      {dollar_sign}stripeClient->createCustomer({dollar_sign}user, {dollar_sign}token);
    } else {
      {dollar_sign}stripeClient->updateCustomerCard({dollar_sign}user, {dollar_sign}token);
    }
    {dollar_sign}cart = ->get('shopping_cart');
    foreach (->getProducts() as ) {
      {dollar_sign}stripeClient->createInvoiceItem(
        {dollar_sign}product->getPrice() * 100,
        {dollar_sign}user,
        {dollar_sign}product->getName()
      );
    }
    {dollar_sign}stripeClient->createInvoice({dollar_sign}user, true);
  }
}
... lines 107 - 108