All we need to do is create a Subscription - via Stripe's API - if they have a plan in their cart. Before we create the Invoice, add if {dollar_sign}cart->getSubscriptionPlan(). Next, open StripeClient: we've designed this class to hold all Stripe API setup and interactions. Add a new method: createSubscription() and give it a User argument and a SubscriptionPlan argument that the User wants to subscribe to.
Now, go back to the Stripe API docs, steal the code that creates a Subscription, and paste it here. Set that to a new  variable. For the customer, use {dollar_sign}user->getStripeCustomerId() to get the id for this user. For the plan, just {dollar_sign}plan->getPlanId(). Return the {dollar_sign}subscription at the bottom.

To use this in the controller, use the {dollar_sign}stripeClient variable we setup earlier: {dollar_sign}stripeClient->createSubscription() and pass it the current {dollar_sign}user variable and then {dollar_sign}cart->getSubscriptionPlan().
And that's all you need to create a subscription!