And, time to save this to the database! Since we're using Doctrine in Symfony, we need the EntityManager object to do this. I'll use dependency injection: add an EntityManager argument to the __construct() method, and set it on a new {dollar_sign}em property.
For Symfony users, this service is using auto-wiring. So because I type-hinted this with EntityManager, Symfony will automatically know to pass that as an argument.

Finally, at the bottom, add {dollar_sign}this->em->persist({dollar_sign}subscription) and {dollar_sign}this->em->flush({dollar_sign}subscription) to save just the Subscription.

With all that setup, go back to OrderController to call this method. To do that, we need the \Stripe\Subscription object. Fortunately, the createSubscription method returns this:
76 lines  src/AppBundle/StripeClient.php
... lines 1 - 8
class StripeClient
{
... lines 11 - 65
  public function createSubscription(User {dollar_sign}user, SubscriptionPlan {dollar_sign}plan)
  {
    {dollar_sign}subscription = \Stripe\Subscription::create(array(
      'customer' => {dollar_sign}user->getStripeCustomerId(),
      'plan' => {dollar_sign}plan->getPlanId()
    ));
    return {dollar_sign}subscription;
  }
}

So add {dollar_sign}stripeSubscription = in front of that line. Then, add {dollar_sign}this->get('subscription_helper')->addSubscriptionToUser() passing it {dollar_sign}stripeSubscription and the currently-logged-in {dollar_sign}user.

Phew! That may have seemed like a lot, but ultimately, this line just makes sure that there is a subscription row in our table that's associated with this user and up-to-date with the subscription and plan IDs.

Let's go try it out. Add a new subscription to your cart, fill out the fake credit card information and hit checkout. No errors! To the account page! Yes! The subscription is active! Our database is up-to-date.