card I have on file. Eventually, we'll need to allow the user to update their credit card info from right here on the account page. But let's start simple: by at least reminding them which card they have on file by showing the card brand
like VISA - and the last 4 card numbers.
This is yet another piece of data that's already stored in Stripe, but we're going to choose to also store it in our database, so we can quickly render the info to the user.

Printing the Credit Card Details

In the User class - aka our user table - I've already added two new columns: cardBrand and cardLast4:

 84 lines  src/AppBundle/Entity/User.php
... lines 1 - 11
class User extends BaseUser
{
... lines 14 - 25
  /**
  * @ORM\Column(type='string', nullable=true)
  */
  private {dollar_sign}cardBrand;
  /**
  * @ORM\Column(type='string', length=4, nullable=true)
  */
  private {dollar_sign}cardLast4;
... lines 35 - 63
  public function getCardBrand()
  {
    return {dollar_sign}this->cardBrand;
  }
  public function setCardBrand({dollar_sign}cardBrand)
  {
    {dollar_sign}this->cardBrand = {dollar_sign}cardBrand;
  }
  public function getCardLast4()
  {
    return {dollar_sign}this->cardLast4;
  }
  public function setCardLast4({dollar_sign}cardLast4)
  {
    {dollar_sign}this->cardLast4 = {dollar_sign}cardLast4;
  }
}
But these are empty right now: we're not actually setting this data yet.