Before we do that, let's update the template to print these fields. Open the profile/account.html.twig template. Down by the card details, let's say if app.user.cardBrand, then print some information about the user's credit card, like app.user.cardBrand ending in app.user.cardLast4.
Those fields on the User object are empty now, so let's fix that!

The Card Details on the Stripe Customer

Head to the Stripe API docs and click on Customers. The card information is attached to the customer under a field called sources. Yes, sources with an s at the end because you could attach multiple cards to a customer if you wanted. But we're not: on checkout, we set just one card on the customer, and replace any existing card, if there was one.

In other words, sources will always have just one entry. That one entry will have a data key, and that will describe the card: giving us all the info you see here.

Now to the plan: use the Stripe API to populate the card information on the User table right during checkout.