Back on the account page, our customers would love us if we would show them when they will be billed next. In other words: when will my subscription renew?
Open the Subscription class - aka the subscription table. I've already added a billingPeriodEndsAt column:
102 lines  src/AppBundle/Entity/Subscription.php
... lines 1 - 10
class Subscription
{
... lines 13 - 40
  /**
  * @ORM\Column(type='datetime', nullable=true)
  */
  private {dollar_sign}billingPeriodEndsAt;
... lines 45 - 86
  /**
  * @return \DateTime
  */
  public function getBillingPeriodEndsAt()
  {
    return {dollar_sign}this->billingPeriodEndsAt;
  }
... lines 94 - 100
}
All we need to do is set this when the subscription is first created. Ok, we also need to update it each month when the subscription is renewed - but we'll talk about that later with webhooks.

The Subscription is created - or retrieved - right here in SubscriptionHelper. This is the spot to set that date.

And hey! This method is passed a \Stripe\Subscription object:

 72 lines  src/AppBundle/Subscription/SubscriptionHelper.php
... lines 1 - 8
class SubscriptionHelper
{
... lines 11 - 45
  public function addSubscriptionToUser(\Stripe\Subscription {dollar_sign}stripeSubscription, User {dollar_sign}user)
  {
... lines 48 - 60
    }
... lines 62 - 70
}
Let's check out the API docs to see what that gives us. Oh yes, it has a current_period_end property, which is a UNIX timestamp. Bingo!