To celebrate, open the account.html.twig template. For 'Next Billing At', add if app.user.subscription - so if they have a subscription - then print app.user.subscription.billingPeriodEndsAt|date('F jS') to format the date nicely.
OK team! Refresh that page! The 'Next Billing At' is... wrong! August 9th! That's today! But no worries, that's just because the field is blank in the database, so it's using today. To really test if this is working, we need to checkout with a new subscription.

Now, in real life, you probably won't allow your users to buy multiple subscriptions. Afterall, we're only storing info in the database about one Subscription, per user. But, for testing, it's really handy to be able to checkout over and over again.

The checkout worked! Click 'Account'. Yes! There is the correct date: September 9th, one month from today. And the VISA card ends in 4242.

Alright: the informational part of the account page is done. But, the user still needs to be able to do some pretty important stuff, like cancelling their subscription - yes, this does happen, it's nothing personal - and updating their credit card. Let's get to it.