The updateCustomerCard() method, however, retrieves the customer... but gets lazy and doesn't return it. Fix that with return {dollar_sign}customer.
Back in OrderController, we've got the \Stripe\Customer... so we're mega dangerous! But instead of updating the fields on User right here, let's do it in SubscriptionHelper. Add a new public function updateCardDetails() method with a User object that should be updated and the \Stripe\Customer object that's associated with it.

Now, this is pretty easy: {dollar_sign}cardDetails = {dollar_sign}stripeCustomer->sources->data[0]. Then, {dollar_sign}user->setCardBrand({dollar_sign}cardDetails) - go cheat with the Stripe API - the fields we want are brand and last4. So, {dollar_sign}cardDetails->brand. And {dollar_sign}user->setCardLast4({dollar_sign}cardDetails->last4). Save only the user to the database with the classic {dollar_sign}this->em->persist({dollar_sign}user) and {dollar_sign}this->em->flush({dollar_sign}user).

Finally, call that method! {dollar_sign}this->get('subscription_helper')->updateCardDetails() and pass it  and {dollar_sign}stripeCustomer.

No matter how you checkout, we're going to make sure your card details are updated in our database!

Before we try it out and prove how awesome we are, I want to add one more thing: I want to be able to tell the user when they will be billed next.