Bad news: eventually, someone will want to cancel their subscription to your amazing, awesome service. So sad. But when that happens, let's make it as smooth as possible. Remember: happy customers!
Like everything we do, cancelling a subscription has two parts. First we need to cancel it inside of Stripe and second, we need to update our database, so we know that this user no longer has a subscription.

Start by adding a cancel button to the account page. In account.html.twig, let's move the h1 down a bit.
Next, add a form with method='POST' and make this float right.
We don't actually need a form, but now we can put a button inside and this will POST up to our server. I don't always do this right, but since this action will change something on the server, it's best done with a POST request. Add a few classes for styling and say 'Cancel Subscription'.

I still need to set the action attribute to some URL... but we need to create that endpoint first!