Open ProfileController. This file renders the account page, but we're also going to put code in here to handle some other things on this page, like cancelling a subscription and updating your credit card.
Create a new public function cancelSubscriptionAction(). Give this a URL: @Route('/profile/subscription/cancel') and a name: account_subscription_cancel.
And, since we'll POST here, we might as well require a POST with @Method - hit tab to autocomplete and add the use statement - then POST.
With the endpoint setup, copy the route name and go back into the template. Update action, with path() then paste the route.

And we are setup!