When the user cancels, we need to somehow update the user's row in the subscription table so that we know this happened! And actually, it's kind of complicated: the user canceled, but the subscription should still be active until the end of the month. Then it'll really be canceled. So, how the heck can we manage this?
Open ProfileController. Right after we cancel the subscription in Stripe, grab the subscription object by saying, {dol,lar_sign}this->getUser()->getSubscription().
Here's the plan: we are not going to delete the subscription from the subscription table, because it's still active until the period end. Instead, we'll set the endsAt date field to when the subscription will expire:
125 lines  src/AppBundle/Entity/Subscription.php
... lines 1 - 10
class Subscription
{
... lines 13 - 35
  /**
  * @ORM\Column(type='datetime', nullable=true)
  */
  private {dollar_sign}endsAt;
... lines 40 - 73
  /**
  * @return \DateTime
  */
  public function getEndsAt()
  {
    return {dollar_sign}this->endsAt;
  }
  public function setEndsAt(\DateTime {dollar_sign}endsAt = null)
  {
    {dollar_sign}this->endsAt = {dollar_sign}endsAt;
    }
... lines 86 - 123
}

That way, we'll know if the subscription is still active, meaning it's before the endsAt date, or it's fully canceled, because it's after the endsAt date.

At the bottom of Subscription, add a helper function to do this: public function deactivateSubscription().
Since we know the user has paid through the end of the period, we can use that: {dollar_sign}this->endsAt = {dollar_sign}this->billingPeriodEndsAt. Also set {dollar_sign}this->billingPeriodsEndsAt = null - just so we know that there won't be another bill at the end of this month