To deactivate the subscription in the controller, it's as easy as saying ->deactivateSubscription() and then saving it to the database with the standard persist() and flush() Doctrine code.
And that should do it! Let's give this guy a try. Go to the account page, then press the new 'Cancel Subscription' button. Ok, looks good! Check the customer page in the Stripe dashboard. Yes! The most recent subscription - the one we're dealing with in our code - is active but will cancel at the end of the month.

But if you look at the Account page, everything here still looks 'Active'. We updated the endsAt field on the subscription, but our code in this template isn't smart enough... yet.