Open account.html.twig. Hmm, I need an easy way to know whether or not a subscription is active, and if it is active, whether or not it's in this canceled state.
To help with this, let's create two methods inside the Subscription class. First, public function isActive().
Meaning: does the user still have an active subscription, even if it will cancel at the month's end? So, if {dollar_sign}this->endsAt === null, then the subscription is definitely active. OR, {dollar_sign}this->endsAt is greater than right now, new \DateTime() meaning the subscription is canceled, but is ending in the future.

The second method we need is public function isCanceled() Meaning: if the subscription is active, has the user actually canceled it or not? This will simply be, return ->endsAt !== null