Ok, back to the account template! This time, to be heros!
First, that 'Cancel Subscription' button should only be there if the user has an active subscription. No problem! Add if app.user.hasActiveSubscription().
But even here, if the user has already canceled their subscription, we don't want to keep showing them this button. Add another if: if app.user.subscription.isCancelled().

Then add a little 'TODO' to add a re-activate button. If they've cancelled, they might remember how cool your service is and want to come back LATER and reactivate! In the else, show them the Cancel button.

Finish up the endif and the other endif. And actually, copy these first two lines: we need to re-use them further below. In the section that tells us whether or not we have an active subscription, we now have three states: 'active', 'active but canceled,' and 'none.' Replace the old if statement with the two that you just copied. If the subscription is canceled, add label-warning and say Canceled. Else, we know it's active. If a user doesn't have any type of active subscription, keep the 'none' from before.
Finally, copy just the first if statement and scroll down to 'Next Billing at'. We should only show the next billing period if the user has an active subscription, not just if they have a related subscription object, because it could be canceled. Paste the if statement over this one.

Finally, do the same thing down below for the credit card: I don't want to confuse someone by showing them credit card information when they don't have a subscription. Phew! Ok, refresh! It's beautiful! There's our todo for the reactivate button and the subscription is canceled. But wait! We don't want the 'Next Billing at' and credit card information to show up.

Ah, that's my bad! The hasActiveSubscription() returns true even if the user already cancelled it.