Expired Subscriptions Cannot be Reactivated Let's get to work in the controller. Like everything, this will have two parts. First, we need to reactivate the subscription in Stripe and second, we need to update our database. For the first part, fetch the trusty StripeClient service object with {dollar_sign}stripeClient = {dollar_sign}this->get('stripe_client'). Next, open StripeClient class. Add a new public function reactivateSubscription(). It will need a User argument whose subscription we should reactivate.
As the Stripe docs mentioned, we can only reactivate a subscription that has not been fully canceled. If today is beyond the period end, then the user will need to create an entirely new subscription. That's why we only show the button in our template during this period.

But just in case, add an 'if' statement: if ! {dollar_sign}user->hasActiveSubscription(), then we'll throw a new exception with the text:

Subscriptions can only be reactivated if the subscription has not actually ended.

Nothing should hit that code, but now we'll know if something does.