To reactivate the Subscription, we first need to fetch it. In the Stripe API docs, find 'Retrieve a Subscription.' Every object can be fetched using the same retrieve method. Copy this. Then, add, {dollar_sign}subscription = and paste. Replace the subscription ID with {dollar_sign}user->getSubscription()->getStripeSubscriptionId().
And remember, if any API call to Stripe fails - like because this is an invalid subscription ID - the library will throw an exception. So we don't need to add extra code to check if that subscription was found.

Finally, reactivate the subscription by setting its plan property equal to the original plan ID, which is {dollar_sign}user->getSubscription()->getStripePlanId().
Then, send the details to Stripe with {dollar_sign}subscription->save(). And just in case, return the {dollar_sign}subscription.
Back in ProfileController, reactivate the subscription with, {dollar_sign}stripeClient->reactivateSubscription(->getUser())
And we are done on the Stripe side.