In ProfileController, add a {dollar_sign}stripeSubscription = in front of the {dollar_sign}stripeClient call. Below that, use {dollar_sign}this->get('subscription_helper')->addSubscriptionToUser() and pass it {dollar_sign}stripeSubscription and the current user.
And that is everything!

Give your user a happy flash message and redirect back to the profile page.

I think we're ready to try this! Go back and refresh the profile. Press reactivate and... our 'cancel subscription' button is back, 'active' is back, 'next billing period' is back and 'credit card' is back. In Stripe, the customer's most recent subscription also became active again. Oh man, this is kind of fun to play with: cancel, reactivate, cancel, reactivate. The system is solid.