Problem 2: Canceling within 1 Hour of Renewal But there's one other, weirder, but similar problem. At the end of the month, 1 hour before charging the user, Stripe creates the invoice. It then waits 1 hour, and tries to charge the user for the first time. So, if your user cancels within that hour, then we also need to fully cancel that subscription to prevent its invoice from being paid.
This is a little trickier: we basically need to see if the user is canceling within that one hour window. To figure that out, create a new variable called {dollar_sign}currentPeriodEnd and set that to a \new DateTime() with the @ symbol and {dollar_sign}subscription->current_period_end.

This converts that timestamp into a \DateTime object.

Now, if  < new \DateTime('+1 hour'), then this means that we're probably in that window and should set {dollar_sign}cancelAtPeriodEnd = false.

An easy way of thinking of this is, if the user is pretty close to the end of their period, then canceling now versus at period end, is almost the same. So, we'll just be careful.

But for this to work, your server's timezone needs to be set to UTC, which is the timezone used by the timestamps sent back from Stripe. If you're not sure, you could give yourself some more breathing room, but fully-canceling anyone's subscription that is within one day of the period end.