Eventually, our customers will need to update the credit card that we… have stored. And on surface, this is pretty easy: the card is just a property on the Stripe customer called source. And basically, we need to nearly duplicate the checkout process: show the user a card form, exchange their card info for a Stripe token, submit that token, and then save it on the user as their new card.
Ok, let's rock! Step 1: add an 'Update Card' button to the account page and make it re-use the same card form we already built for checkout. Because, ya know, reusing code is awesome!

In account.html.twig, find the credit card section. Add a button, give it some classes for styling and a js-open-credit-card-form class.

In a second, we'll attach some JavaScript to this.

Next, find the col-xs-6. This is the right side of the page, and I've kept it empty until now just so we can show the form here. Add a div with a JS class so we can hide/show this element: js-update-card-wrapper. Make it display: none; by default. Inside, add a cute header.