I want to re-use the entire checkout form right here. Fortunately, we already isolated that into its own template: _cardForm.html.twig.
In account.html.twig, use the Twig include() function to bring that in.

Ok! Let's hide/show this form whenever the user clicks the 'Update Card' button. At the top of the file, override the block called javascripts, and call endblock. Inside, call the parent() function.

In this project, any JS we put here will be included on the page.

Add a script tag and a very simple document.ready() block. Inside of that, find the .js-open-credit-card-form element and on click, create a callback function. Start with the normal e.preventDefault().

Now, find the other wrapper element, which is js-update-card-wrapper. Call slideToggle() on that to show/hide it.

So, fairly easy stuff.

Well, maybe we should see if it works first. Refresh! Ah, it doesn't! Huge error:

Variable error' does not exist in _cardForm.html.twig at line 56
Hmm, checkout that template. Ah yes, on the checkout page, after we submit, if there was an error, we set this variable and render it here. For now, we don't have any errors. In account.html.twig, we could pass an error variable to the include() call. But, we could also do it in ProfileController::accountAction(). Add error set to null.
Refresh and click 'Update Card'. We are in business!