Sharing Card JavaScript The second problem is pretty obvious if you fill out the card number field on checkout, and then compare it with the profile page! On the checkout page, we included a lot of JavaScript that does cool stuff like format this field. And, much more importantly, the JS is also responsible for sending the credit card information to Stripe, fetching the token, putting it in the form, and submitting it. We definitely still need that.
Ok, how can we reuse the JavaScript? In checkout.html.twig, we just inlined all of our JS right in the template. That's not great, but since this isn't a course about JavaScript, let's solve this as easily as possible. Copy all of the JS and create a new template called _creditCardFormJavaScript.html.twig inside the order/ directory. Paste this there.

Now, in checkout, include that template! Copy that and include the same thing in account.html.twig at the top of the javascripts block.

Ok, refresh and hope for the best! Ah, another missing variable: stripe_public_key:

Variable stripe_public_key' does not exist in _creditCardFormJavaScript.html.twig at line 5
We're printing this in the middle of our JS.
And the variable comes from OrderController. Copy that line, open ProfileController, and paste it there.
Now the page works! And - at the very least - the JS formatting is rocking.

Frontend stuff is done: let's submit this and update the user's card in Stripe.