Now, what about our database? Do we store any information about the credit card? Actually, we do! In the User class, we store cardLast4 and cardBrand. With the new card, this stuff probably changed!
But wait, we've got this handled too guys! Open SubscriptionHelper and check out the handy updateCardDetails() method.
Just pass it the User and \Stripe\Customer and it'll make sure those fields are set.

In ProfileController, call this: {dollar_sign}this->get('subscription_helper')->updateCardDetails() passing {dollar_sign}user and {dollar_sign}stripeCustomer.
That's it! Add a success message so that everyone feels happy and joyful, and redirect back to the profile page.
Time to try it! Refresh and put in the fake credit card info. But use a different expiration: 10/25. Hit 'Update Card'. Ok, it looks like it worked. Refresh the Customer page in Stripe. The expiration was 10/20 and now it's 10/25. Card update successful!

But, we still need to handle one more case: when the card update fails.